Remove extra , and rename helpers/ to apis/
